chore(ui): move components only used in conversation detail page into conversations/[id]/ directory
